refactor: move telemetry interfaces and types to api file
